refactor: 리뷰 상세조회 api 수정 및 타입 전역 분리
