+++ src/pages/ReviewReplyPage/components/ReceiverItem/index.tsx
@@ -1,8 +1,8 @@
 import { MouseEvent } from 'react'
 import { useLocation } from 'react-router-dom'
 import { Profile } from '@/components'
-import { Receiver } from '@/apis/hooks/useGetReviewForCreator'
 import { CheckInTheCircleIcon } from '@/assets/icons'
+import { Receiver } from '@/types'
 
 interface ReceiverItemProps {
   receiver: Receiver
