+++ src/pages/ReviewReplyPage/components/ReplyCategory/ReplyHexa/index.tsx
@@ -1,11 +1,11 @@
 import { useEffect } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useLocation } from 'react-router-dom'
-import { QuestionOption } from '@/apis/hooks/useGetReviewForCreator'
 import {
   ReviewReplyStartType,
   ReviewReplyEditType,
 } from '@/pages/ReviewReplyPage/types'
+import { QuestionOption } from '@/types'
 
 interface ReplyHexaProps {
   receiverIndex: number
