+++ src/pages/ReviewReplyPage/components/ReviewReplyEdit/index.tsx
@@ -3,11 +3,11 @@ import { FormProvider, useForm, useFieldArray } from 'react-hook-form'
 import { useNavigate, useLocation } from 'react-router-dom'
 import {
   useEditResponse,
-  useGetReviewForCreator,
+  useGetReviewForParticipation,
   useUser,
   useGetResponseByResponserForReviewReplyPage,
 } from '@/apis/hooks'
-import { Question } from '@/apis/hooks/useGetReviewForCreator'
+import { Question } from '@/types'
 import { ReviewReplyEditType } from '../../types'
 import ReviewReply from './ReviewReply'
 
@@ -22,7 +22,7 @@ const ReviewReplyEdit = () => {
     reviewId,
     responserId: user?.id as number,
   })
-  const { data: reviewData } = useGetReviewForCreator({ id: reviewId })
+  const { data: reviewData } = useGetReviewForParticipation({ id: reviewId })
   const { mutate: editResponse } = useEditResponse()
   const { title, questions } = reviewData
