+++ src/pages/ReviewReplyPage/components/ReviewReplyStart/ReviewReply.tsx
@@ -1,5 +1,5 @@
 import { useFormContext } from 'react-hook-form'
-import { Data } from '@/apis/hooks/useGetReviewForCreator'
+import { ReviewDetailedData } from '@/types'
 import { QuestionItem, ReceiverItem } from '..'
 import {
   useHandleReceiver,
@@ -11,7 +11,7 @@ import { ReviewReplyStartType } from '../../types'
 import Questions from '../Questions'
 
 interface ReviewReplyProps {
-  reviewData: Data
+  reviewData: ReviewDetailedData
   handleSubmit?: () => void
 }
