+++ src/pages/ReviewReplyPage/hooks/useClickNextButton.ts
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from 'react'
-import { Question, Receiver } from '@/apis/hooks/useGetReviewForCreator'
+import { Question, Receiver } from '@/types'
 
 interface UseClickNextButtonProps {
   questions: Question[]
