+++ src/pages/ReviewReplyPage/hooks/useHandleQuestion.ts
@@ -1,7 +1,7 @@
 import { useState, MouseEvent } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useLocation } from 'react-router-dom'
-import { Question } from '@/apis/hooks/useGetReviewForCreator'
+import { Question } from '@/types'
 import { ReviewReplyStartType } from '../types'
 
 interface UseHandleQuestionProps {
