+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './reviews'
+export * from './reviewDetailed'
